refactor: move integration command handlers to commands/integration.py
Extracts integration_app, all @integration_app.command handlers, and
helper functions (_read_integration_json, _write_integration_json,
_remove_integration_json, _normalize_script_type, _resolve_script_type,
_update_init_options_for_integration) from __init__.py into the focused
commands/integration.py module. __init__.py re-exports the public
symbols via a clean import block.
